+++ common/act/procinst/where.go
@@ -10,28 +10,28 @@ import (
 )
 
 // ID filters vertices based on their ID field.
-func ID(id int) predicate.ProcInst {
+func ID(id int64) predicate.ProcInst {
 	return predicate.ProcInst(func(s *sql.Selector) {
 		s.Where(sql.EQ(s.C(FieldID), id))
 	})
 }
 
 // IDEQ applies the EQ predicate on the ID field.
-func IDEQ(id int) predicate.ProcInst {
+func IDEQ(id int64) predicate.ProcInst {
 	return predicate.ProcInst(func(s *sql.Selector) {
 		s.Where(sql.EQ(s.C(FieldID), id))
 	})
 }
 
 // IDNEQ applies the NEQ predicate on the ID field.
-func IDNEQ(id int) predicate.ProcInst {
+func IDNEQ(id int64) predicate.ProcInst {
 	return predicate.ProcInst(func(s *sql.Selector) {
 		s.Where(sql.NEQ(s.C(FieldID), id))
 	})
 }
 
 // IDIn applies the In predicate on the ID field.
-func IDIn(ids ...int) predicate.ProcInst {
+func IDIn(ids ...int64) predicate.ProcInst {
 	return predicate.ProcInst(func(s *sql.Selector) {
 		v := make([]interface{}, len(ids))
 		for i := range v {
@@ -42,7 +42,7 @@ func IDIn(ids ...int) predicate.ProcInst {
 }
 
 // IDNotIn applies the NotIn predicate on the ID field.
-func IDNotIn(ids ...int) predicate.ProcInst {
+func IDNotIn(ids ...int64) predicate.ProcInst {
 	return predicate.ProcInst(func(s *sql.Selector) {
 		v := make([]interface{}, len(ids))
 		for i := range v {
@@ -53,28 +53,28 @@ func IDNotIn(ids ...int) predicate.ProcInst {
 }
 
 // IDGT applies the GT predicate on the ID field.
-func IDGT(id int) predicate.ProcInst {
+func IDGT(id int64) predicate.ProcInst {
 	return predicate.ProcInst(func(s *sql.Selector) {
 		s.Where(sql.GT(s.C(FieldID), id))
 	})
 }
 
 // IDGTE applies the GTE predicate on the ID field.
-func IDGTE(id int) predicate.ProcInst {
+func IDGTE(id int64) predicate.ProcInst {
 	return predicate.ProcInst(func(s *sql.Selector) {
 		s.Where(sql.GTE(s.C(FieldID), id))
 	})
 }
 
 // IDLT applies the LT predicate on the ID field.
-func IDLT(id int) predicate.ProcInst {
+func IDLT(id int64) predicate.ProcInst {
 	return predicate.ProcInst(func(s *sql.Selector) {
 		s.Where(sql.LT(s.C(FieldID), id))
 	})
 }
 
 // IDLTE applies the LTE predicate on the ID field.
-func IDLTE(id int) predicate.ProcInst {
+func IDLTE(id int64) predicate.ProcInst {
 	return predicate.ProcInst(func(s *sql.Selector) {
 		s.Where(sql.LTE(s.C(FieldID), id))
 	})
